Handling negative values and test rewriting
In the last stage, the tests, which had so far had negative values, were changed and a specially dedicated test was created.

In this step, the changes specified in the requirements have been implemented. The occurrence of negative numbers results in the occurrence of an exception and the listing of these values.

SeeAlso: # 5